requirements.txt
sounddevice==0.5.1
numpy==1.26.4
cryptography==43.0.0
supabase==2.10.0
pillow==10.4.0
pyinstaller==6.11.1
